memorylocker: Replace with new implementation
